feat(wallet): ajuste de precisão com BigDecimal e escala
- Atualizado todos os campos financeiros para usar BigDecimal com setScale
- saldoFiat e saldoTrade: scale 8
- quantidade de moedas: scale 4
- valores de operação e preço médio: scale 8
- corrigido problemas de arredondamento e conversão
- padronizado RoundingMode.HALF_UP
- atualizado também Notificacao, Carteira, Operacao e CarteiraMoeda

+++ src/main/java/br/com/oinkvest/controller/NotificationController.java
@@ -1,13 +1,20 @@
 package br.com.oinkvest.controller;
 
-import br.com.oinkvest.model.Notificacao;
-import br.com.oinkvest.model.Usuario;
-import br.com.oinkvest.service.NotificationService;
-import br.com.oinkvest.config.UsuarioDetails;
+import java.math.BigDecimal;
+
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+import br.com.oinkvest.config.UsuarioDetails;
+import br.com.oinkvest.model.Notificacao;
+import br.com.oinkvest.model.Usuario;
+import br.com.oinkvest.service.NotificationService;
 
 @Controller
 @RequestMapping("/notifications")
@@ -32,7 +39,7 @@ public class NotificationController {
     @PostMapping
     public String criarAlerta(@AuthenticationPrincipal UsuarioDetails usuarioDetails,
                               @RequestParam String moeda,
-                              @RequestParam double precoAlvo,
+                              @RequestParam BigDecimal precoAlvo,
                               @RequestParam Notificacao.Condicao condicao) {
 
         Usuario usuario = usuarioDetails.getUsuario();
